refactor(coder): Phase 4 - Extract todo collection to separate file
Phase 4.2 Complete:
- Created todo_collection.go with requestTodoList, createTodoCollectionToolProvider, processTodosFromEffect
- Removed ~120 lines from plan_review.go
- Improves separation of concerns between plan review and todo collection

Phase 4.1 and 4.3 Assessed and Deferred:
- Container helpers (~120 lines) are already well-organized, extraction deferred
- Container fallback logic (setup.go:203-218) is only 2 levels deep, clear as-is

Updated CODER_CODE_REVIEW.md:
- Marked Phase 4 complete
- Updated executive summary and top risks
